backend_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-bigquery>=3.13.0
google-cloud-firestore>=2.13.1
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.2